# src/books/dto/update-book.dto.ts
import { PartialType, OmitType } from "@nestjs/mapped-types";
import { CreateBookDto } from "./create-book.dto";

export class UpdateBookDto extends PartialType(
  OmitType(CreateBookDto, ["authorId"] as const)
) {}